services/altair/requirements.txt
pyarrow
pandas
altair
panel
streamlit
# postgres
psycopg2-binary
SQLAlchemy
